Add files via upload
The program counts number of words, symbols with spaces and symbols without spaces. If a price for translation a provided, also count price for every file and sums to a total price. User can set price for 1 word, for X symbols with spaces or X symbols without spaces. Results are shown as table which can be exported to *.xlsx file. The program works with document types *.txt and *.docx. 
Programming language: Python 3.5